Домашнее задание 3 добавлен класс ResultTest по замечаниям из прошлого pr
