+++ Homework 03/src/main/java/ru/otus/spring082022/Beloborodov03/service/QuestionServiceImpl.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.otus.spring082022.Beloborodov03.appconfig.AppProps;
 import ru.otus.spring082022.Beloborodov03.domain.Question;
+import ru.otus.spring082022.Beloborodov03.domain.ResultTest;
 import ru.otus.spring082022.Beloborodov03.domain.Student;
 import ru.otus.spring082022.Beloborodov03.repositary.QuestionDAO;
 
@@ -29,11 +30,12 @@ public class QuestionServiceImpl implements QuestionService {
         String firstName = inoutService.inStringWithPrompt(true, "hello.firstname");
         String lastName = inoutService.inStringWithPrompt(true, "hello.lastname");
         Student student = new Student(firstName, lastName);
+        ResultTest resultTest = new ResultTest(student);
 
         questions.forEach((q) -> {
             inoutService.outString(true, "result.question");
             inoutService.outStringn(false, q.getQuestionText());
-            student.addAnswer(inoutService.inStringWithPrompt(true, "result.youranswer"));
+            resultTest.addAnswer(inoutService.inStringWithPrompt(true, "result.youranswer"));
         });
         int rightAnswers = 0;
         inoutService.outString(true, "result.student");
@@ -43,7 +45,7 @@ public class QuestionServiceImpl implements QuestionService {
             inoutService.outStringn(false, questions.get(i).getQuestionText());
             inoutService.outString(true, "result.rightanswer");
             inoutService.outString(false, questions.get(i).getRightAnswerText());
-            String answer = student.getAnswer(i);
+            String answer = resultTest.getAnswer(i);
             if (answer != null) {
                 inoutService.outString(true, "result.youranswer");
                 inoutService.outStringn(false, answer);
